float -> double, instead of checking normalization nozmalize orientation and add vec_math and vectors ariphmetic

+++ srcs/parsing/utils/helpers.c
@@ -6,7 +6,7 @@
 /*   By: mansargs <mansargs@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/15 14:37:41 by mansargs          #+#    #+#             */
-/*   Updated: 2025/12/15 16:05:29 by mansargs         ###   ########.fr       */
+/*   Updated: 2026/01/05 19:47:57 by mansargs         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ bool	all_spaces(const char *line)
 	return (true);
 }
 
-bool	out_of_range(float num, float range_min, float range_max)
+bool	out_of_range(double num, double range_min, double range_max)
 {
 	return (num < range_min || num > range_max);
 }
